% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckCountryCodes}
\alias{CheckCountryCodes}
\title{Country codes are used consistently}
\usage{
CheckCountryCodes()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks all origin and asylum codes and ensures that all used are valid
}
\examples{
{
}
}
\author{
Edgar Scrase
}
