% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckDemographicAgeUnknown}
\alias{CheckDemographicAgeUnknown}
\title{Demographic disaggregation by age and sex has a small percentage of unknown age cohorts}
\usage{
CheckDemographicAgeUnknown()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks the % of the demographic data (where dissaggregation by age
 and sex is available) for which the male and female unknown columns are 
 relatively high (a smaller percentage is better).
}
\examples{
{
}
}
\author{
Edgar Scrase
}
