% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckDemographicAggregationType}
\alias{CheckDemographicAggregationType}
\title{Aggregation type correctly describes the demographic disaggregation recorded in each row}
\usage{
CheckDemographicAggregationType()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
There are four types of valid demographic data aggregations - :
 'Detailed' where there is full disaggregation using the detailed age cohorts for adults,
  , 'M/F and 18-59' where only the 18-59 age cohort is available for adults,
   'M/F' where only disaggregation by sex is available and 
    'Total' where no disaggregation is available.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
