% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckDemographicCoverageAgeSex}
\alias{CheckDemographicCoverageAgeSex}
\title{Demographic coverage by country of asylum and population type - sex and age disaggregation}
\usage{
CheckDemographicCoverageAgeSex()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks the completeness of coverage by age and sex. 
Try to extend the coverage of the available demographic data through new data sources,
 statistical modelling or estimation.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
