% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckDemographicTotals}
\alias{CheckDemographicTotals}
\title{Demographic totals are consistent with the other tables}
\usage{
CheckDemographicTotals()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Compares the figures for each population type and countries of 
origin and asylum that are presented in the demographic table with those
 in each of the source tables.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
