% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckGeneralMissingBasis}
\alias{CheckGeneralMissingBasis}
\title{All data should include a basis and source}
\usage{
CheckGeneralMissingBasis()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks for missing basis and source in the relevant source tables.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
