% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckHostCommunity}
\alias{CheckHostCommunity}
\title{Total reported host community population figures should be consistent 
with the assisted figures}
\usage{
CheckHostCommunity()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks that the total and assisted host community figures are equivalent -
 this is the most typical case for reporting on host communities
}
\examples{
{
}
}
\author{
Edgar Scrase
}
