% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckIDPDataIDMC}
\alias{CheckIDPDataIDMC}
\title{IDMC comparison for stock and new arrivals}
\usage{
CheckIDPDataIDMC()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Compares the IDMC and UNHCR IDP figures, highlighting significant differences.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
