% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckRefugeeBasis}
\alias{CheckRefugeeBasis}
\title{Refugee Basis should be from registration or census statistics}
\usage{
CheckRefugeeBasis()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Where possible, refugee figures should be based on government or 
census data rather than estimates.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
