% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckRefugeeReturns}
\alias{CheckRefugeeReturns}
\title{The number of refugee returns reported by the country of origin is 
consistent with the country(ies) of former asylum}
\usage{
CheckRefugeeReturns()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Review together the differences in the number of returns recorded 
by the country of origin and the country(ies) of asylum.  
Small differences are completely acceptable, but large differences should be rare.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
