% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckRSDAfrica}
\alias{CheckRSDAfrica}
\title{Countries in Africa should typically not recognize individuals under "other"}
\usage{
CheckRSDAfrica()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
While there are exceptions, in most cases recognitions through RSD
 should be under UNHCR's mandate rather than complementary protection in countries in Africa.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
