% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckRSDOverTime}
\alias{CheckRSDOverTime}
\title{Checks the consistency of the procedure type in the RSD table over time}
\usage{
CheckRSDOverTime()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
If a country reports type of RSD procedure Government one year and
 type of procedure as Joint the following year
 If a country reports Government or Joint cases in the past and then reports only UNHCR cases
}
\examples{
{
}
}
\author{
Edgar Scrase
}
