% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckSignificantChange}
\alias{CheckSignificantChange}
\title{Change over time is within acceptable boundaries.}
\usage{
CheckSignificantChange()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks that the percentage change from previous years 
(+- 5-10% for refugee population of 100 or more) is within acceptable boundaries.
  The check is conducted by population type and countries of origin and asylum.
}
\examples{
{
}
}
\author{
Edgar Scrase
}
