% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckStatelessOrigin}
\alias{CheckStatelessOrigin}
\title{Displaced stateless correctly reported by their specific countries of
 former habitual residence}
\usage{
CheckStatelessOrigin()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
Checks the percentage of the origin in the Stateless table that is 
STA/UKN versus specific countries of former habitual residence for 
displaced STA (e.g. the Rohingya)
}
\examples{
{
}
}
\author{
Edgar Scrase
}
