% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckVDAInOOC}
\alias{CheckVDAInOOC}
\title{VDA should no longer be included in the OOC table}
\usage{
CheckVDAInOOC()
}
\arguments{
\item{ctr}{country or list of countries - using alpha3 code - to pull using popdata package}

\item{sqafYear}{numeric value for the year of the check - for instance 2020}

\item{sqafIsASR}{boolean indicating if it is Annual or mid year report TRUE}

\item{Threshold_1}{}

\item{Threshold_2}{}

\item{Threshold_4}{}
}
\value{

}
\description{
This checks that Venezuelans Displaced Abroad are no longer 
included in the OOC table, as there is now a separate table to include these
 populations.  There could potentially be special cases here so not an absolute error
}
\examples{
{
}
}
\author{
Edgar Scrase
}
